Refactoring code of module tags
