+++ angular-spa-ui/js/search-app/tags-module/create-tag-ctrl.js
@@ -1,4 +1,3 @@
-
 "use strict";
 
 module.exports = function($scope, tagService) {
