+++ angular-spa-ui/js/search-app/tags-module/tags-ctrl.js
@@ -2,8 +2,8 @@
 
 module.exports = function($scope, tagService) {
 
-    // Implementation config of main controller Tags
-    // Property of object (get static params)
+    // Implementation main controller of 'module tags'
+    // Get property of current scope
     $scope.titlePage = tagService.titlePage;
     $scope.titleCreateTag = tagService.titleCreateTag;
     $scope.titleResultTagList = tagService.titleResultTagList;
