+++ angular-spa-ui/js/search-app/tags-module/tags-list-dir.js
@@ -1,0 +1,18 @@
+"use strict";
+
+module.exports = function() {
+
+    var tagsList = {
+        restrict: 'E',
+        transclude: true,
+        scope: {
+
+        },
+        controller: 'tagsListContrller',
+        templateUrl: '/listTagsTemplate.html',
+        replace: true
+    };
+
+    return tagsList;
+
+};
